+++ ground/src/main/java/com/google/android/ground/persistence/local/LocalDataStoreModule.kt
@@ -18,6 +18,8 @@ package com.google.android.ground.persistence.local
 import com.google.android.ground.persistence.local.room.LocalDatabase
 import com.google.android.ground.persistence.local.room.RoomLocalDataStore
 import com.google.android.ground.persistence.local.room.dao.*
+import com.google.android.ground.persistence.local.room.stores.*
+import com.google.android.ground.persistence.local.stores.*
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
@@ -30,6 +32,14 @@ import javax.inject.Singleton
 abstract class LocalDataStoreModule {
 
   @Binds @Singleton abstract fun localDataStore(ds: RoomLocalDataStore): LocalDataStore
+  @Binds
+  @Singleton
+  abstract fun locationOfInterestStore(store: RoomLocationOfInterestStore): LocationOfInterestStore
+  @Binds @Singleton abstract fun offlineAreaStore(store: RoomOfflineAreaStore): OfflineAreaStore
+  @Binds @Singleton abstract fun submissionStore(store: RoomSubmissionStore): SubmissionStore
+  @Binds @Singleton abstract fun surveyStore(store: RoomSurveyStore): SurveyStore
+  @Binds @Singleton abstract fun tileSetStore(store: RoomTileSetStore): TileSetStore
+  @Binds @Singleton abstract fun userStore(store: RoomUserStore): UserStore
 
   companion object {
     @Provides
