+++ ground/src/main/java/com/google/android/ground/persistence/local/stores/Store.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.android.ground.persistence.local.stores
+
+import java8.util.function.Supplier
+import timber.log.Timber
+
+/**
+ * The shared type of classes responsible for managing access to Ground's local-only stored data.
+ * These classes return Ground model objects that represent an object as its currently stored in the
+ * local database.
+ *
+ * When data needs to be persisted and synchronized in both the local and remote data storage use
+ * [MutationStore], which has additional methods for handling synchronization.
+ */
+sealed interface Store<T> {
+  fun <R> logAndSkip(supplier: Supplier<R>): Iterable<R> =
+    try {
+      setOf(supplier.get())
+    } catch (e: RuntimeException) {
+      Timber.d(e.message)
+      emptySet()
+    }
+}
